Fix dry-run flag

+++ kubernetes/stop_greenbox.sh
@@ -6,6 +6,7 @@
 KUBE_CONTEXT=$1
 CROMWELL_URL=$2
 VAULT_ENV=$3
+DRY_RUN=$4
 VAULT_TOKEN_FILE=${VAULT_TOKEN_FILE:-"$HOME/.vault-token"}
 
 if [ -z $KUBE_CONTEXT ]; then
@@ -17,24 +18,28 @@ elif [ -z $CROMWELL_URL ]; then
 elif [ -z $VAULT_ENV ]; then
     echo -e "\nYou must specify a vault environment for getting the CaaS service account key"
     error=1
+elif [ -z $DRY_RUN ]; then
+    DRY_RUN = "false"
 fi
-
 if [ $error -eq 1 ]; then
     echo -e "\nUsage: bash stop_greenbox.sh KUBE_CONTEXT CROMWELL_URL VAULT_ENV\n"
     exit 1
 fi
 
-
 kubectl config use-context ${KUBE_CONTEXT}
 
 # Delete ingress rule for Lira to stop receiving notifications
 echo "Delete Lira ingress"
-kubectl delete ingress listener
+if [ $DRY_RUN == "false" ]; then
+    kubectl delete ingress listener
+fi
 
 # Bring down Falcon to stop releasing workflows
 # TODO: Uncomment when Falcon is deployed
-#echo "Delete Falcon deployment"
-#kubectl delete deployment falcon
+# echo "Delete Falcon deployment"
+# if [ $DRY_RUN == 'false' ]; then
+# kubectl delete deployment falcon
+# fi
 
 CAAS_KEY_FILE="caas_key.json"
 docker run -i --rm -e VAULT_TOKEN=$(cat $VAULT_TOKEN_FILE) broadinstitute/dsde-toolbox vault read \
@@ -43,4 +48,4 @@ docker run -i --rm -e VAULT_TOKEN=$(cat $VAULT_TOKEN_FILE) broadinstitute/dsde-t
         secret/dsde/mint/$VAULT_ENV/listener/caas-${VAULT_ENV}-key.json > $CAAS_KEY_FILE
 
 # Abort all on-hold and running workflows
-python -m abort_workflows --cromwell_url ${CROMWELL_URL} --caas_key ${CAAS_KEY_FILE}
+python -m abort_workflows --cromwell_url ${CROMWELL_URL} --caas_key ${CAAS_KEY_FILE} --dry_run ${DRY_RUN}
